Simplify GHA. (#122)
* Simplify GHA.

* Maybe it'll work back on Mac?

rstanarm is having issues.

* Remove rstanarm.

We don't actually use it and it's hard to install. Simplify!

* Specify ubuntu version.

Darnit, rstanarm is imported by tidyposterior. So instead specify the same ubuntu version that rstanarm uses in their checks.

* Remove unused packages.

We don't use these to render the book.

* Go back to latest.

rstanarm caused problems. Let's see if it will work on latest now (and test caching while we're at it).